Fix animation for walking left
Introduce a new AnimationState for standing still and fix the
transitions between states. Also add an atlas index for the new
State that tracks the index of the previous state so the player
stops facing in the right direction.
Add constants for storing the start indices of the respective
animations in the atlas.